pyproject.toml
[project]
name = "nonencm"
version = "0.2.2"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "loguru",
    "mutagen",
    "noneprompt",
    "pillow",
    "pyncm>=1.8.1",
    "pyyaml",
    "qqmusic-api-python>=0.3.6",
    "segno",
]

[project.scripts]
nonencm = "app.__main__:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project.optional-dependencies]
build = [
    "pyinstaller>=6.0.0",
]
pil-utils = [
    "pil-utils>=0.2.0",
    "typing_extensions>=4.8.0",
]

[tool.setuptools.packages.find]
include = ["app*"]
exclude = ["old*", "logs*"]
